+++ src/MenuScreen.cpp
@@ -52,7 +52,7 @@ void MenuScreen::init() {
 void MenuScreen::handleInput(sf::RenderWindow& window, const sf::Event& event) {
     if (event.type == sf::Event::MouseButtonPressed) {
         if (event.mouseButton.button == sf::Mouse::Left) {
-            sf::Vector2f mousePos(event.mouseButton.x, event.mouseButton.y);
+            sf::Vector2f mousePos(static_cast<float>(event.mouseButton.x), static_cast<float>(event.mouseButton.y));
             
             for (auto& btn : m_buttons) {
                 if (btn.rect.getGlobalBounds().contains(mousePos)) {
@@ -70,7 +70,7 @@ void MenuScreen::handleInput(sf::RenderWindow& window, const sf::Event& event) {
     
     // Hover effect
     if (event.type == sf::Event::MouseMoved) {
-        sf::Vector2f mousePos(event.mouseMove.x, event.mouseMove.y);
+        sf::Vector2f mousePos(static_cast<float>(event.mouseMove.x), static_cast<float>(event.mouseMove.y));
         for (auto& btn : m_buttons) {
             if (btn.rect.getGlobalBounds().contains(mousePos)) {
                 btn.rect.setFillColor(sf::Color(150, 150, 150));
